ipa-uuid: support base64-encoded autogenerated uuid
Active Directory LDAP schema includes objectGUID attribute which is
encoded as an octet string. For containers in Global Catalog we need to
specify objectGUID value in each object. For those objects that describe
a structure of Global Catalog, it is easier to autogenerate them
directly.

Add support to produce base64-encoded autogenerated uuid to ipa-uuid
plugin. If configuration has 'ipaUuidEncode: TRUE' attribute value, it
will be created as base64-encoded one. This is incompatible with
prefixed UUIDs and should be only used for octet strings.