Global Catalog: fix convert-schema
The convert-schema tool is creating a schema for Global Catalog
based on the definitions from Active Directory Schema. By default
it is using the syntax 1.3.6.1.4.1.1466.115.121.1.26 (IA5String)
for attributes, unless a mapping is specified.

Active Directory is using the syntax 2.5.5.1 (Distinguished Name Syntax)
for member, manager, or distinguishedName attributes and the tool
currently converts them to IA5String instead of DN.
The patch adds a mapping to convert 2.5.5.1 to 1.3.6.1.4.1.1466.115.121.1.12
instead (DN), so that member, manager are seen as DNs.