GlobalCatalog: bump 389-ds version for AD filter rewriter for ObjectCategory
With version 1.4.3.6, 389-ds provides a rewriter for filters using
objectCategory. The rewriters allows to translate filters such as
"(objectCategory=foo)" into
"(objectcategory=cn=foo,cn=schema,cn=configuration,$base)"

This feature is needed by the Global Catalog to mimic AD behavior.

With the version bump, the schema is also more strict and fixes are
needed for the convert-schema tool:
- the attribute syntax 2.5.5.17 (Sid) that is translated to
1.3.6.1.4.1.1466.115.121.1.40 (octet string) cannot use caseExactMatch
and caseExactOrderingMatch for equality and ordering, but should use
octetStringMatch and octetStringOrderingMatch instead.
- the objectclass applicationEntity is now defined in 389-ds schema
- the  objectclass dMD conflicts with dmd defined in 389-ds schema. Replace
with dMD-ad