ipa-adtrust install: use 'installer' context
Currently ipa-adtrust-install is using the 'install' context
but this means that services started/stopped inside this context
with SystemdService.start() or SystemdService.stop() won't
update the service list in /var/run/ipa/services.list
(as there is a check if context == ipactl or installer).

As a result, they won't be stopped by the first call made to
ipactl stop right after they have been installed.

Change ipa-adtrust-install context to 'installer' instead of 'install',
as this fixes the above issue.