Harden uninstallation of Global Catalog
When ipa-server-install --uninstall is called on a replica
without Global Catalog, the code tries to remove the service
alias for the GC but fails if the service entry is absent.
This can happen if ipa server-del was run before uninstalling
the replica, but should not prevent the uninstall from running.