gcinstance: preconfigure service principal aliases for GC as expected by MS-DRSR
AD DCs expect RPC service principal name to exist for the machine
account of the domain controller they talk to. Since in FreeIPA case
Kerberos principals for host/ and cifs/ are separate, we need to add
RPC/... alias to cifs/... principal.

There are other service principal names for GC that are expected to
exist according to MS-DRSR 2.2.3 and 2.2.4 sections. Refactor GC
instance code to include all of them.

Signed-off-by: Alexander Bokovoy <abokovoy@redhat.com>