GC: uninstall does not remove all the service aliases
During the uninstallation of global catalog, the aliases added
to ldap/$FQDN@$DOMAIN need to be removed. One of them
is not properly templated and remains after the uninstallation:
E3514235-4B06-11D1-AB04-00C04FC2DCD2/$DOMAINGUID_TEXT/$DOMAIN

This happens because of a wrong initialization of $DOMAINGUID_TEXT
using ipantflatname instead of ipantdomainguid.

The installer should also handle gracefully when it tries to
add an alias that is already present.