Global Catalog: remove DNS record only when last instance is removed
During the GC uninstallation, all the GC-related DNS records are
deleted, even though another master may be providing the GC role.
Delete the records only when the last instance of GC is removed.

If the deletion fails for an unexpected reason, do not raise the
exception but continue the uninstallation after logging the error.

Fixes: https://github.com/abbra/freeipa/issues/63