GC: remove DNS records for last server uninstall
The uninstallation of the last server with GC role can
happen after various code paths:
- server-del + uninstall
- direct uninstall

In order to detect that the last GC is being uninstalled, the
code relies on the output of server-role-find(GC, enabled).
The current server is the last GC in the following cases:
- EmptyResult exception
- 0 result
- 1 result corresponding to the current server.

In all the above cases, the uninstaller must remove GC DNS records.
Fixes: https://pagure.io/freeipa/issue/8758

Signed-off-by: Florence Blanc-Renaud <flo@redhat.com>