WIP: subid: allow ID overrides from Default Trust View to own subid
subid entries need a DN to denote an owner. This DN can point to any
LDAP entry. Allow to use ID overrides from the Default Trust View for
this purpose, effectively allowing to add subordinate IDs from the IPA
subid ID range to trusted Active Directory users.

In order to write the ownership information, trusted user account must
be a member of any roles that grant 'Subordinate ID Selfservice Users'
privilege. This can be done, for example, by creating a group and adding
the ID override as a member of that group and adding a group as a member
of 'Subordinate ID Selfservice User' role.

Signed-off-by: Alexander Bokovoy <abokovoy@redhat.com>