ipa-sam: fix TDO objects with within forest flag
When trusted domain object already exists in the database with
LSA_TRUST_ATTRIBUTE_WITHIN_FOREST, it needs to be corrected to use
LSA_TRUST_ATTRIBUTE_FOREST_TRANSITIVE before passing it to Samba.

We don't need to update the values in the LDAP object though.

Fixes: https://pagure.io/freeipa/issue/9892
Related: https://pagure.io/freeipa/issue/9852

Signed-off-by: Alexander Bokovoy <abokovoy@redhat.com>